% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRedis.R
\docType{package}
\name{easyRedis}
\alias{easyRedis}
\alias{easyRedis-package}
\alias{easyRedis-package}
\title{easyRedis: Helper Package for Redis Caches}
\description{
easyRedis: Helper Package for Redis Caches
}
\section{Functions}{

Some functions
}

