% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_redis_api2.R
\name{ErInit}
\alias{ErInit}
\title{ErInit: init a EasyRedis object}
\usage{
ErInit(host = "localhost", port = 6379, password = NULL)
}
\arguments{
\item{host}{redis host.}

\item{port}{redis port}

\item{password}{redis password}
}
\value{
a EasyRedis object
}
\description{
ErInit: init a EasyRedis object
}
\examples{
er = EasyRedis::ErInit()
x = "apple"
er$qSet(x)
er$get("x") # "apple"

}
