% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_redis_api2.R
\name{getEnv}
\alias{getEnv}
\title{getEnv get env value.}
\usage{
getEnv(x)
}
\arguments{
\item{x}{env value name}
}
\value{
the value. if value == "" then return NULL
}
\description{
getEnv get env value.
}
