% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_redis_api2.R
\name{init}
\alias{init}
\title{init a EasyRedis object}
\usage{
init(host = NULL, port = 6379, password = NULL)
}
\arguments{
\item{password}{}
}
\value{
a EasyRedis object
}
\description{
init a EasyRedis object
}
\examples{
er = EasyRedis::init()
x = "apple"
er$qset(x)
er$get("x") # "apple"

}
