% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageOptions.R
\name{redis_options}
\alias{redis_options}
\title{Package Options}
\usage{
redis_options(..., ifnotfound = NULL)
}
\arguments{
\item{...}{either a list of character strings (names of options to get) or
a named list of options to set}

\item{ifnotfound}{default return value if option is not set}
}
\value{
For \code{...} being an unnamed list of character strings, the
    function returns a list of options. If the resulting list is of length
    one, the list will be flattene, i.e. only the first item will be
    returned. If not parameter will be supplied, the complete list of
    options will be returned.

    For \code{...} being a names list of options to be set, the function
    invisibly returns the original values (if set) of the set options.
}
\description{
Get and set package options.
}
