% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setConfig.R
\name{redis_setConfig}
\alias{redis_setConfig}
\title{Set Config Settings}
\usage{
redis_setConfig(key, ..., .list)
}
\arguments{
\item{key}{character, key to identify config settings}

\item{...}{named arguments for config settings}

\item{.list}{list of config settings, see details}
}
\value{
a list of configuration settings
}
\description{
The function sets config entries with a given key in the Redis cache.
}
\details{
The config settings may either be passed using \code{...} argument
    or \code{.list}. If \code{.list} is provided, any other parameter in
    \code{...} will be omitted.

    If \code{key} does not exist, a new key will be generated. The ID will,
    if not provided within the list of config settings, generated using
    \code{\link[uuid]{UUIDgenerate}()}.

    If \code{key} exists and config settings contain an ID, an error will be
    raised if the key's ID and the config settings list ID do not match.
}
