% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRedis.R
\docType{package}
\name{easyRedis}
\alias{easyRedis}
\alias{easyRedis-package}
\alias{easyRedis-package}
\title{easyRedis: Helper Package for Redis Caches}
\description{
The package provides various function for working with Redis data bases
}
\section{Functions}{

\describe{
   \item{\code{\link{redis_options}}}{get or set options for \pkg{easyRedis}
       package}
   \item{\code{\link{redis_connect}}, \code{\link{redis_close}}}{connect to
       Redis data base or close the connection}
   \item{\code{\link{redis_loadConfig}}}{loads configuration files to
       Redis data base}
   \item{\code{\link{redis_getConfig}}}{get configuration settings from
       Redis data base}
   \item{\code{\link{redis_logger}}}{define a logger to publish messages to
       a Redis data base}

}
}

