% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfig.R
\name{redis_getConfig}
\alias{redis_getConfig}
\title{Get Configuration Data from Redis Cache}
\usage{
redis_getConfig(key)
}
\arguments{
\item{key}{character string, key to identify config data in Redis cache}
}
\value{
an object deparsed from JSON string that is stored in Redis cache
    with key \code{key}.

    The return value has two attributes
    \describe{
        \item{\code{id}}{the unique identifier of the config data}
        \item{\code{key}}{the Redis cache key}
    }

    If \code{key} is not found, \code{NULL} is returned.
}
\description{
Get Configuration Data from Redis Cache
}
