% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadConfig.R
\name{redis_loadConfig}
\alias{redis_loadConfig}
\title{Load Config-Files to Redis Cache}
\usage{
redis_loadConfig(file, ..., append = FALSE)
}
\arguments{
\item{file}{character vector of file names}

\item{...}{additional paramter passed to \code{\link{redis_connect}} in
case of no open Redis connection}

\item{append}{logical, if \code{TRUE} config settings will be appended to
existing settings (if any). Duplicate settings will be overwritten.}
}
\value{
a list of same length as `file` containing configuration settings;
    if `file` is of length 1, only the first list element will be returned
}
\description{
The function loads config data, stored as JSON-strings in config files to
the Redis cache.
}
\section{Data Format}{

All config files must contain data in JSON format. Basic elements of the
data structure at top level are
\describe{
    \item{\code{key}}{\emph{mandatory}, unique key to store and access data
        in the Redis cache}
    \item{\code{value}}{\emph{mandatory}, data object containing a list of
        key-value pairs. This can be any list of named data. There is one
        \emph{mandatory} element: \code{id}. This is a unique identifier of
        the data. It will be matched to \code{key} and no key must be
        assigned to one or more \code{id}s (e.g. from multiple
        config-files)}
    \item{\code{auto_load}}{\emph{optional}, an auto load obejct, i.e. a
        list of key-value pairs, see Auto Load section for details.}
}

For loading a minimal config file, see Examples.
}

\section{Auto Load}{

The \code{auto_load} section of the configuration file contains a named
list of code to create configuration settings dynamically. Code can be
any valid expression e.g. \preformatted{"date": "Sys.Date()"}
or, using \code{::} to locate a function within a package,
\preformatted{"date": "lubridate::today()"}
to set config value \code{date} to the current date (using
\code{\link[base]{Sys.Date}()}).

Code can also be a file name with an optional expression (separated by a
colon) to be evaluated within the context of this file. If no expression
is given, variable \code{value} as defined within the file, will be returned.

Relative paths for file names are given with respect to the location of
the config file.
}

\examples{
\dontrun{
## show template for minimal configuration file
file <- system.file('templates/minimal.cfg', package = 'easyRedis')
cat(stringr::str_c(readLines(file), collapse = '\\n'))

## *** SET UP AN OPEN REDIS-CONNECTION FIRST! ***
config <- redis_loadConfig(file)

key <- attr(config, 'key')

redis_getConfig(key)

## *** CLEAR REDIS KEY! ***
rredis::redisDelete(key)
}

}
\references{
see \url{https://en.wikipedia.org/wiki/JSON} for details on
    JSON file format
}
